hostname-util: treat hostname == FALLBACK_HOSTNAME same as unset
If kernel hostname is equal to FALLBACK_HOSTNAME (the compiled-in
default), treat it the same as if it was empty or equal to "(none)".

This way, a distribution may configure a default hostname both in the
kernel and the systemd it ships, while at the same time allowing it to
be overridden via os-release if specific builds (e. g. live CD) call for
a different default.