core/cgroup: add "unset" and "default" values for ZSwapWriteback
Instead of treating ZSwapWriteback (memory.zswap.writeback) as a bool
and eagerly initializing it for each cgroup, let kernel inherit the
values and introduce a dedicated "unset" state where we let the kernel
do its thing.

Additionally, introduce a fourth "default" state which corresponds to
the `-1` value of the memory.zswap.writeback attribute, meaning "use
global default" (this is not yet in the kernel; will be submitted
shortly).